Extract opening of the ssh session
